add exception handling

+++ src/main/java/by/shabunya/eventmanager/dao/impl/EventDAO.java
@@ -2,8 +2,8 @@ package by.shabunya.eventmanager.dao.impl;
 
 import by.shabunya.eventmanager.dao.GenericDAO;
 import by.shabunya.eventmanager.entity.Event;
+import by.shabunya.eventmanager.exception.EventNotFoundException;
 import org.hibernate.Session;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
@@ -14,7 +14,6 @@ public class EventDAO implements GenericDAO<Event> {
 
     private final EntityManager entityManager;
 
-    @Autowired
     public EventDAO(EntityManager entityManager) {
         this.entityManager = entityManager;
     }
@@ -27,7 +26,10 @@ public class EventDAO implements GenericDAO<Event> {
 
     @Override
     public Event get(Long id) {
-        return entityManager.unwrap(Session.class).get(Event.class, id);
+        Event event = entityManager.unwrap(Session.class).get(Event.class, id);
+        if (event == null)
+            throw new EventNotFoundException("No event with id " + id, id);
+        return event;
     }
 
     @Override
@@ -39,6 +41,10 @@ public class EventDAO implements GenericDAO<Event> {
     public void update(Event event, Long id) {
         Session session = entityManager.unwrap(Session.class);
         Event eventToUpdate = session.get(Event.class, id);
+
+        if (eventToUpdate == null)
+            throw new EventNotFoundException("No event with id " + id, id);
+
         session.evict(eventToUpdate);
         eventToUpdate.setTheme(event.getTheme());
         eventToUpdate.setDescription(event.getDescription());
@@ -52,6 +58,10 @@ public class EventDAO implements GenericDAO<Event> {
     public void delete(Long id) {
         Session session = entityManager.unwrap(Session.class);
         Event eventToDelete = session.get(Event.class, id);
+
+        if (eventToDelete == null)
+            throw new EventNotFoundException("No event with id " + id, id);
+
         session.delete(eventToDelete);
     }
 }
